requirements.txt
opencv-contrib-python
pyserial
keras
tensorflow
Image
pandas
sklearn